# api/routes/openapi/openapi.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "info": {
    "title": "IndexCards API",
    "version": "1.2.0",
    "description": "Tabroom.com data & operational API",
    "termsOfService": "https://www.speechanddebate.org/terms-conditions/",
    "license": {
      "name": "Copyright 2014-2021, National Speech & Debate Association",
      "identifier": "RPL-1.5"
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ],
...
  "paths": {
...
    "/rest/tourns/{tournId}/rounds/{roundId}": {
      "get": {
        "summary": "Returns round information given an ID if it is public",
        "operationId": "getRound",
        "responses": {
          "200": {
            "description": "Object of Round with public information on it",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "invite",
          "public",
          "schematics",
          "rounds",
          "pairings"
        ],
        "description": "GET /rest/tourns/{tournId}/rounds/{roundId} is undocumented. Need to add .openapi to handler",
        "parameters": [
          {
            "$ref": "#/components/parameters/tournId"
          },
          {
            "$ref": "#/components/parameters/roundId"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference that identifies the problem type."
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code."
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation of the error."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "description": "A URI reference to the specific occurrence."
          }
        }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Unexpected Error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "error": {
                "summary": "500 Internal Server Error",
                "value": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "An unexpected error occurred.",
                  "instance": "/api/v1/example"
                }
              }
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized - authentication failed or was not provided.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "unauthorized": {
                "summary": "401 Unauthorized",
                "value": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "You are not authorized to access this resource."
                }
              }
            }
          }
        }
      }
    },
    "parameters": {
      "tournId": {
        "name": "tournId",
        "in": "path",
        "description": "The ID or webname of the tournament to retrieve.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "roundId": {
        "name": "roundId",
        "in": "path",
        "description": "The ID or webname of the round to retrieve.",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "securitySchemes": {
      "extApiKey": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "x-tabroom-cookie"
      }
    }
  }
}